Move CSRF to another path
